wip: update ot withVisible HOC
