+++ guides/Architecture.md
@@ -30,6 +30,20 @@ not documentation should explain why the convention has been overridden.
   `classNames` that will be computed along with the base component classes.
 - Tests for the component should cover the exposed component API.
 
+### Active components
+
+#### Responsibilities
+
+_These responsibilites inform what components should be required to manage and
+prevents creep_
+
+- `active-container-factory` generates container components for active elements.
+  Handles coordination of controlled and uncontrolled active usage, and provides
+  the active provider to set the active context for that element.
+- `ActiveProvider` provides context for passing+accessing active state and
+  update methods at any level.
+- `withActive` passes the values and change handlers with a context consumer.
+
 ### Computing props
 
 Every library component can have default props configured using the
